Explain the model

+++ ann/utils.jl
@@ -1,7 +1,7 @@
 using DataFrames
 
 # bound values in an array
-function bound(a::Vector{Int64}; min = 0, max = 256)
+function bound(a::Vector{Float64}; min = 0.0, max = 256.0)
   underflow = find(x -> x < min, a)
   overflow = find(x -> x > max, a)
   a[underflow] = min
